Using Policy Based Data Structure (Indexed Set)
Note that if it were not the case that all elements of the input array were distinct, then this code would be incorrect since Tree<int> would remove duplicates. Instead, we would use an indexed set of pairs (Tree<pair<int, int>>), where the first element of each pair would denote the value while the second would denote the position of the value in the array.